+++ app/src/main/java/com/kuit/ourmenu/ui/navigator/MainNavController.kt
@@ -118,7 +118,7 @@ class MainNavController(
     }
 
     // Menu Info
-    fun navigateToMenuInfo(menuId: Int) {
+    fun navigateToMenuInfo(menuId: Long) {
         navController.navigateToMenuInfo(menuId)
     }
